app/requirements.txt
streamlit
openai
requests
python-dotenv
numpy==1.26.2
farm-haystack[all]